requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
redis==5.0.3
argon2-cffi==23.1.0
pyotp==2.9.0
python-jose[cryptography]==3.3.0
boto3==1.35.54
